Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@6036 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/karts/kart_model.cpp
@@ -461,7 +461,7 @@ void KartModel::update(float rotation, float steer, const float suspension[4])
     else if(steer<0.0f) frame = m_animation_frame[AF_STRAIGHT]
                               + ( (m_animation_frame[AF_STRAIGHT] 
                                         -m_animation_frame[AF_LEFT]   )*steer);
-    else                frame = m_animation_frame[AF_STRAIGHT];
+    else                frame = (float)m_animation_frame[AF_STRAIGHT];
 
     m_animated_node->setCurrentFrame(frame);
 }   // update
